Added a filters data structure
