+++ src/const.js
@@ -1,0 +1,18 @@
+
+const CARDS_COUNT = 12;
+const CARDS_PER_CLICK = 5;
+const COMMENTS_COUNT = 50;
+
+const FilterType = {
+  ALL: 'All movies',
+  WATCHLIST: 'Watchlist',
+  HISTORY: 'History',
+  FAVORITES: 'Favorites',
+};
+
+export {
+  CARDS_COUNT,
+  CARDS_PER_CLICK,
+  COMMENTS_COUNT,
+  FilterType,
+};
