+++ src/mock/filter.js
@@ -1,0 +1,8 @@
+import {filter} from '../utils/filter';
+
+export const generateFilter = (cards) => Object.entries(filter)
+  .map(([filterName, filterTask]) => ({
+    name: filterName,
+    count: filterTask(cards).length
+  }),
+  );
