+++ src/view/filter-view.js
@@ -1,16 +1,23 @@
 import AbstractView from '../framework/view/abstract-view';
 
-const createTemplate = () => (`
+const createTemplate = (filter) => `
   <nav class="main-navigation">
-    <a href="#all" class="main-navigation__item main-navigation__item--active">All movies</a>
-    <a href="#watchlist" class="main-navigation__item">Watchlist <span class="main-navigation__item-count">13</span></a>
-    <a href="#history" class="main-navigation__item">History <span class="main-navigation__item-count">4</span></a>
-    <a href="#favorites" class="main-navigation__item">Favorites <span class="main-navigation__item-count">8</span></a>
-  </nav>
-`);
+    ${filter.reduce((template, item) => {
+    template += `
+      <a href="#${item.name.replace(/\s+/g, '').toLowerCase()}" class="main-navigation__item">${item.name}<span class="main-navigation__item-count">${item.count}</span></a>`;
+    return template;}, '')
+}
+  </nav>`;
 
-export default class FilterView extends AbstractView{
+export default class FilterView extends AbstractView {
+  #filter = null;
+
+  constructor(filter) {
+    super();
+    this.#filter = filter;
+  }
+
   get template() {
-    return createTemplate();
+    return createTemplate(this.#filter);
   }
 }
